requirements.txt
langchain
langchain-google-genai
pillow
streamlit
python-dotenv
langchain_community
unstructured
langchain_chroma
langchain_openai